+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+- oihana\enums\FilterOption
+
 ----
 
 ## [1.0.0] - 2025-08-13